# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "ACROSS Server",
        "summary": "Astrophysics Cross-Observatory Science Support (ACROSS)",
        "description": "Server providing tools and utilities for various NASA missions to aid in coordination of large observation efforts.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/role/": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Read roles",
                "description": "Read many roles.",
                "operationId": "get_roles",
                "responses": {
                    "200": {
                        "description": "A list of roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Role"
                                    },
                                    "type": "array",
                                    "title": "Response 200 Get Roles"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The role does not exist."
                    }
                }
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Create",
                "operationId": "create_role",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "title": "Response Create Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/role/{role_id}": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Read a role",
                "description": "Read a role by role ID.",
                "operationId": "get_role",
                "parameters": [
                    {
                        "name": "role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Role Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/group/{group_id}/role": {
            "get": {
                "tags": [
                    "Group Role"
                ],
                "summary": "Read group roles",
                "description": "Read many group roles.",
                "operationId": "get_group_roles",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of group roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GroupRoleRead"
                                    },
                                    "title": "Response 200 Get Group Roles"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Group Role"
                ],
                "summary": "Create a group role",
                "description": "Create a group role with a set of permissions in a group that you control",
                "operationId": "create_group_role",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GroupRoleCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created group role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupRoleRead"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/group/{group_id}/role/{group_role_id}": {
            "get": {
                "tags": [
                    "Group Role"
                ],
                "summary": "Read a group role",
                "description": "Read a group role by role ID.",
                "operationId": "get_group_role",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    },
                    {
                        "name": "group_role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Group Role Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested group role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/across_server__routes__group__role__schemas__GroupRole"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Group Role"
                ],
                "summary": "Update a group role",
                "description": "Update a group role and its permissions in a group that you control",
                "operationId": "update_group_role",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Group Role Id"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GroupRoleCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated group role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupRoleCreate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Group Role"
                ],
                "summary": "Delete a group role",
                "description": "Delete a group role by role ID.",
                "operationId": "delete_group_role",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    },
                    {
                        "name": "group_role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Group Role Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Group role successfully deleted"
                    },
                    "404": {
                        "description": "The group or role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/group/{group_id}/user/{user_id}/role/{group_role_id}": {
            "put": {
                "tags": [
                    "Group Role"
                ],
                "summary": "Assign a group role to a user",
                "description": "Assign a group role by id to a user id within that group.",
                "operationId": "assign_to_user",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Group Role Id"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Group role successfully assigned to user"
                    },
                    "404": {
                        "description": "The group or role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Group Role"
                ],
                "summary": "Remove a group role from a user",
                "description": "Remove a group role by id from a user id within that group.",
                "operationId": "remove_from_user",
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "parameters": [
                    {
                        "name": "group_role_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Group Role Id"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "UUID of the group"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Group role successfully removed from user"
                    },
                    "404": {
                        "description": "The group or role does not exist."
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GroupRoleCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "permission_ids": {
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "type": "array",
                        "title": "Permission Ids"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "permission_ids"
                ],
                "title": "GroupRoleCreate"
            },
            "GroupRoleRead": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "permissions": {
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        },
                        "type": "array",
                        "title": "Permissions"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "permissions",
                    "id"
                ],
                "title": "GroupRoleRead"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Permission": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "title": "Permission"
            },
...
            "Role": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "users": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__role__schemas__User"
                        },
                        "type": "array",
                        "title": "Users"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id",
                    "users"
                ],
                "title": "Role"
            },
...
            "RoleCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "permissions": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Permissions"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "permissions"
                ],
                "title": "RoleCreate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
            "across_server__routes__group__role__schemas__GroupRole": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "permissions": {
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        },
                        "type": "array",
                        "title": "Permissions"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "users": {
                        "items": {
                            "$ref": "#/components/schemas/across_server__routes__group__role__schemas__User"
                        },
                        "type": "array",
                        "title": "Users"
                    },
                    "service_accounts": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/across_server__routes__group__role__schemas__ServiceAccount"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Service Accounts"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "permissions",
                    "id",
                    "users",
                    "service_accounts"
                ],
                "title": "GroupRole"
            },
            "across_server__routes__group__role__schemas__ServiceAccount": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "user": {
                        "$ref": "#/components/schemas/across_server__routes__group__role__schemas__User"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "description",
                    "user"
                ],
                "title": "ServiceAccount"
            },
            "across_server__routes__group__role__schemas__User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "username",
                    "email"
                ],
                "title": "User"
            },
...
            "across_server__routes__role__schemas__User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "username",
                    "email"
                ],
                "title": "User"
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "description": "Enter your access token.",
                "scheme": "bearer"
            }
        }
    }
}